# back/src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gonglijeongli API",
    "description": "API Documentation for Gonglijeongli Backend",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Relative URL for current domain"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Dev Server"
    }
  ],
  "paths": {
...
    "/api/news/": {
      "get": {
        "tags": [
          "News"
        ],
        "summary": "뉴스 목록 조회 (페이지네이션)",
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "페이지 번호",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "페이지당 항목 수",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "News"
        ],
        "summary": "새 뉴스 생성",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "뉴스 데이터",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Sample News"
                },
                "content": {
                  "type": "string",
                  "example": "This is the news content"
                },
                "thumbnail": {
                  "type": "string",
                  "example": "https://example.com/image.jpg"
                },
                "date": {
                  "type": "string",
                  "example": "2023-10-31"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "example": "any"
                  },
                  "content": {
                    "example": "any"
                  },
                  "thumbnail": {
                    "example": "any"
                  },
                  "date": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/news/{id}": {
      "get": {
        "tags": [
          "News"
        ],
        "summary": "뉴스 상세 조회",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "뉴스 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "News"
        ],
        "summary": "뉴스 수정",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "뉴스 ID"
          },
          {
            "name": "body",
            "in": "body",
            "description": "수정할 뉴스 데이터",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Updated News"
                },
                "content": {
                  "type": "string",
                  "example": "Updated content"
                },
                "thumbnail": {
                  "type": "string",
                  "example": "https://example.com/updated.jpg"
                },
                "date": {
                  "type": "string",
                  "example": "2023-11-01"
                },
                "likes": {
                  "type": "number",
                  "example": 10
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "example": "any"
                  },
                  "content": {
                    "example": "any"
                  },
                  "thumbnail": {
                    "example": "any"
                  },
                  "date": {
                    "example": "any"
                  },
                  "likes": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "News"
        ],
        "summary": "뉴스 삭제",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "뉴스 ID"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/api/upload/": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "이미지 업로드 (관리자)",
        "description": "",
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/upload/cleanup": {
      "delete": {
        "tags": [
          "Upload"
        ],
        "summary": "사용하지 않는 미참조 이미지 일괄 삭제 (가비지 컬렉션)",
        "description": "",
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    }
  }
}